requirement.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
django-cors-headers==2.4.0
django-filter==2.0.0
django-ipware==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.5
